use api/columns/outline give different page
